+++ lib/zap_message/webhook.rb
@@ -3,6 +3,7 @@
 require 'zap_message/webhook/base_event'
 require 'zap_message/webhook/message_received'
 require 'zap_message/webhook/message_status'
+require 'zap_message/webhook/system_event'
 
 module ZapMessage
   module Webhook